+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 ### 3.0.0-alpha.3
 
 - permission.get: default search with deleted=0
-
+- session.put: added
+- routes/permission: added GET, POST, DELETE
+- mysql(insert, select, update, delete): return just the query
+- 
 
 
 [3.0.0-alpha.3]: https://github.com/NicTool/api/releases/tag/3.0.0-alpha.3