+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### 3.0.0-alpha.3
 
+- routes/permission: added GET, POST, DELETE
 - permission.get: default search with deleted=0
 - session.put: added
-- routes/permission: added GET, POST, DELETE
+- session: store user/group info in cookie (saves DB trips)
 - mysql(insert, select, update, delete): return just the query
-- 
+- lib/group.get: convert booleans
+- lib/user.get: convert booleans
 
 
 [3.0.0-alpha.3]: https://github.com/NicTool/api/releases/tag/3.0.0-alpha.3